# React App Performance Optimization: Identify & Resolve Slowdown